Only catch creator can edit/delete

+++ apirouter.js
@@ -14,6 +14,7 @@ router.get('/catches', function(req, res) {
 router.post('/catches', function(req, res) {
   const newCatch = new Catch({
     pokemon: req.body.pokemon,
+    creator: req.user,
     description: req.body.description,
     lat: req.body.lat,
     lng: req.body.lng
@@ -28,33 +29,51 @@ router.post('/catches', function(req, res) {
 });
 
 router.delete('/catches/:id', function(req, res) {
-  Catch.deleteOne({ _id: req.params.id }, err => {
+  Catch.findById(req.params.id, (err, foundCatch) => {
     if (err) {
       return res.status(404).json({ message: 'not found' });
     }
-    return res.status(200).json({ message: 'success' });
+    if (foundCatch.creator === req.user) {
+      Catch.deleteOne({ _id: req.params.id }, err => {
+        if (err) {
+          return res.status(404).json({ message: 'not found' });
+        }
+        return res.status(200).json({ message: 'success' });
+      });
+    } else {
+      return res.status(403).json({ message: 'not allowed' });
+    }
   });
 });
 
 router.put('/catches/:id', function(req, res) {
-  Catch.findOneAndUpdate(
-    { _id: req.params.id },
-    {
-      $set: {
-        pokemon: req.body.pokemon,
-        description: req.body.description,
-        lat: req.body.lat,
-        lng: req.body.lng
-        //date: req.body.date
-      }
-    },
-    err => {
-      if (err) {
-        return res.status(409).json({ message: 'conflict' });
-      }
-      return res.status(200).json({ message: 'success' });
+  Catch.findById(req.params.id, (err, foundCatch) => {
+    if (err) {
+      return res.status(404).json({ message: 'not found' });
     }
-  );
+    if (foundCatch.creator === req.user) {
+      Catch.findOneAndUpdate(
+        { _id: req.params.id },
+        {
+          $set: {
+            pokemon: req.body.pokemon,
+            description: req.body.description,
+            lat: req.body.lat,
+            lng: req.body.lng
+            //date: req.body.date
+          }
+        },
+        err => {
+          if (err) {
+            return res.status(409).json({ message: 'conflict' });
+          }
+          return res.status(200).json({ message: 'success' });
+        }
+      );
+    } else {
+      return res.status(403).json({ message: 'not allowed' });
+    }
+  });
 });
 
 module.exports = router;
